Fix: JenkinsLauncher instead
